requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.13.0=pypi_0
argon2-cffi=20.1.0=pypi_0
astunparse=1.6.3=pypi_0
async-generator=1.10=pypi_0
attrs=21.2.0=pypi_0
backcall=0.2.0=pypi_0
bleach=3.3.0=pypi_0
brotli=1.0.9=pypi_0
bzip2=1.0.8=h8ffe710_4
ca-certificates=2021.5.30=h5b45459_0
cachetools=4.2.2=pypi_0
certifi=2021.5.30=py39hcbf5309_0
cffi=1.14.5=pypi_0
chardet=4.0.0=pypi_0
click=8.0.1=pypi_0
colorama=0.4.4=pypi_0
colorlover=0.3.0=pypi_0
cycler=0.10.0=pypi_0
dash=1.19.0=pypi_0
dash-bootstrap-components=0.12.2=pypi_0
dash-canvas=0.1.0=pypi_0
dash-core-components=1.15.0=pypi_0
dash-daq=0.5.0=pypi_0
dash-html-components=1.1.2=pypi_0
dash-renderer=1.9.0=pypi_0
dash-table=4.11.2=pypi_0
decorator=4.4.2=pypi_0
defusedxml=0.7.1=pypi_0
entrypoints=0.3=pypi_0
et-xmlfile=1.1.0=pypi_0
face-clustering=0.0.1=pypi_0
flask=1.1.2=pypi_0
flask-compress=1.9.0=pypi_0
flatbuffers=1.12=pypi_0
freetype=2.10.4=h546665d_1
future=0.18.2=pypi_0
gast=0.4.0=pypi_0
google-api-core=1.31.1=pypi_0
google-auth=1.32.1=pypi_0
google-auth-oauthlib=0.4.4=pypi_0
google-cloud-vision=2.4.2=pypi_0
google-pasta=0.2.0=pypi_0
googleapis-common-protos=1.53.0=pypi_0
grpcio=1.34.1=pypi_0
gunicorn=20.0.4=pypi_0
h5py=3.1.0=pypi_0
idna=2.10=pypi_0
imageio=2.9.0=pypi_0
intel-openmp=2021.3.0=h57928b3_3372
ipykernel=5.5.5=pypi_0
ipython=7.24.1=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.6.3=pypi_0
itsdangerous=2.0.1=pypi_0
jbig=2.1=h8d14728_2003
jedi=0.18.0=pypi_0
jinja2=3.0.1=pypi_0
joblib=1.0.1=pyhd8ed1ab_0
jpeg=9d=h8ffe710_0
jsonschema=3.2.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=6.1.12=pypi_0
jupyter-console=6.4.0=pypi_0
jupyter-core=4.7.1=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-widgets=1.0.0=pypi_0
keras=2.4.3=pypi_0
keras-nightly=2.5.0.dev2021032900=pypi_0
keras-preprocessing=1.1.2=pypi_0
keras-vggface=0.6=pypi_0
kiwisolver=1.3.1=pypi_0
lcms2=2.12=h2a16943_0
leptonica=1.78.0=h688788b_4
lerc=2.2.1=h0e60522_0
libarchive=3.5.1=hb45042f_2
libblas=3.9.0=9_mkl
libcblas=3.9.0=9_mkl
libdeflate=1.7=h8ffe710_5
libiconv=1.16=he774522_0
liblapack=3.9.0=9_mkl
libpng=1.6.37=h1d00b33_2
libtiff=4.3.0=h0c97f57_1
libwebp=1.2.0=h57928b3_0
libwebp-base=1.2.0=h8ffe710_2
libxml2=2.9.12=hf5bbc77_0
lz4-c=1.9.3=h8ffe710_0
lzo=2.10=he774522_1000
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.3.4=pypi_0
markupsafe=2.0.1=pypi_0
matplotlib=3.4.2=pypi_0
matplotlib-inline=0.1.2=pypi_0
mistune=0.8.4=pypi_0
mkl=2021.2.0=hb70f87d_389
msys2-conda-epoch=20160418=1
mtcnn=0.1.0=pypi_0
nbclient=0.5.3=pypi_0
nbconvert=6.0.7=pypi_0
nbformat=5.1.3=pypi_0
nest-asyncio=1.5.1=pypi_0
networkx=2.5.1=pypi_0
notebook=6.4.0=pypi_0
numpy=1.19.5=pypi_0
oauthlib=3.1.1=pypi_0
olefile=0.46=pyh9f0ad1d_1
opencv-python=4.5.2.54=pypi_0
openjpeg=2.4.0=hb211442_1
openpyxl=3.0.7=pypi_0
openssl=1.1.1k=h8ffe710_0
opt-einsum=3.3.0=pypi_0
packaging=20.9=pypi_0
pandas=1.2.4=pypi_0
pandocfilters=1.4.3=pypi_0
parso=0.8.2=pypi_0
pickleshare=0.7.5=pypi_0
pillow=8.3.0=py39h916092e_0
pip=21.1.1=py39haa95532_0
plotly=4.14.3=pypi_0
prometheus-client=0.11.0=pypi_0
prompt-toolkit=3.0.19=pypi_0
proto-plus=1.19.0=pypi_0
protobuf=3.17.3=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycontracts=1.8.12=pypi_0
pycparser=2.20=pypi_0
pygments=2.9.0=pypi_0
pyparsing=2.4.7=pypi_0
pyrsistent=0.17.3=pypi_0
pytesseract=0.3.8=pyhd8ed1ab_0
python=3.9.5=h6244533_3
python-dateutil=2.8.1=pypi_0
python_abi=3.9=2_cp39
pytz=2021.1=pypi_0
pywavelets=1.1.1=pypi_0
pywin32=301=pypi_0
pywinpty=1.1.2=pypi_0
pyyaml=5.4.1=pypi_0
pyzmq=22.1.0=pypi_0
qtconsole=5.1.0=pypi_0
qtpy=1.9.0=pypi_0
requests=2.25.1=pypi_0
requests-oauthlib=1.3.0=pypi_0
retrying=1.3.3=pypi_0
rsa=4.7.2=pypi_0
scikit-image=0.18.2=pypi_0
scikit-learn=0.24.2=py39he931e04_0
scipy=1.7.0=py39hc0c34ad_0
send2trash=1.5.0=pypi_0
setuptools=52.0.0=py39haa95532_0
six=1.15.0=pypi_0
sqlite=3.35.4=h2bbff1b_0
tbb=2021.3.0=h2d74725_0
tensorboard=2.5.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
tensorflow=2.5.0=pypi_0
tensorflow-estimator=2.5.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.10.1=pypi_0
tesseract=4.1.1=h8c00e3c_4
testpath=0.5.0=pypi_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tifffile=2021.7.2=pypi_0
tk=8.6.10=h8ffe710_1
tornado=6.1=pypi_0
traitlets=5.0.5=pypi_0
typing-extensions=3.7.4.3=pypi_0
tzdata=2020f=h52ac0ba_0
urllib3=1.26.6=pypi_0
vc=14.2=h21ff451_1
visdcc=0.0.40=pypi_0
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=2.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=pypi_0
wincertstore=0.2=py39h2bbff1b_0
wrapt=1.12.1=pypi_0
xlrd=2.0.1=pypi_0
xz=5.2.5=h62dcd97_1
zlib=1.2.11=h62dcd97_1010
zstd=1.5.0=h6255e5f_0
